Brint bootstrap template closer to current default (#218)
I compared the bootstrap template to the current default generated by CDK and noticed that it was hard to compare due to different formatting and because this one was based on v8 while CDK has moved on to v14. So I went through the changes carefully and ported them to the v14 of the template. Now, if you run a diff between the original bootstrap template and the one provided here, only the modifications required for the bootstrap kit will stand out.